Renames zsh submodules directories for fzf-lua
fzf-lua was indexing the submodules directories and became
a mess. This commit renames the directory to `submodules`
to reduce the chance of a conflict in the future with other
projects where we ignore directories we shouldn't.

Then, we add the pattern to fzf-lua config so the files can
be ignored during searches